% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinvar_data.R
\docType{data}
\name{clinvar_data}
\alias{clinvar_data}
\title{Default ClinVar Data}
\format{
dataframe with 82872 rows and 5 variables:
\describe{
\item{cv_variant_id}{ClinVar variant identifer.}
\item{uniprot_id}{UniProt accession identifier.}
\item{transcript_id}{Ensembl transcript identifier.}
\item{protein_variant}{Protein variant identifier.}
\item{cv_class}{Binary ClinVar class. 0 for benign or 1 for pathogenic.}
}
}
#'\usage{
#'clinvar_data
#'}
\description{
Derived from the supplemental table of the AlphaMissense
\href{https://www.science.org/doi/10.1126/science.adg7492}{[2023]} paper.
}
\examples{
data(clinvar_data)
}
\keyword{datasets}
