% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_tables}
\alias{db_tables}
\alias{db_temporary_table}
\alias{db_range_join}
\alias{db_disconnect}
\title{Manipulate the Database of Missense Mutations}
\usage{
db_tables(db)

db_temporary_table(db, value, to)

db_range_join(db, key, join, to)

db_disconnect(db)
}
\arguments{
\item{db}{a \code{duckdb_connection} object, as returned by \code{am_data()}.}

\item{value}{a \code{data.frame} / \code{tibble} containing data to be placed
in a temporary table, e.g., from a GenomicRanges object to be
used in a range join.}

\item{to}{the character(1) name of the table to be created}

\item{key}{a character(1) table name in \code{db} containing missense
mutation coordinates.}

\item{join}{a character(1) table name in \code{db} containing ranges to
be used for joining with (filtering) \code{key}.}
}
\value{
\code{db_tables()} returns a character vector of database table
names.

\code{db_temporary_table()} returns the temporary table as a
dbplyr \code{tibble}.

\code{db_range_join()} returns \code{to} (the temporary table created
from the join) as a dbplyr tibble.

\code{db_disconnect()} returns \code{FALSE} if the connection has
already been closed or is not valid (via \code{dbIsValid()}) or
\code{TRUE} if disconnection is successful. Values are returned
invisibly.
}
\description{
\code{db_tables()} queries for the names of temporary and
regular tables defined in the database.

\code{db_temporary_table()} creates a temporary (for the
duration of the duckdb connection) table from a tibble.

\code{db_range_join()} performs a range join, finding all
positions in \code{key} within ranges defined by \code{join}. The result
is stored in table \code{to}.

\code{db_disconnect()} disconnects the duckdb database and
shuts down the DuckDB server associated with the
connection. Temporary tables are lost.
}
\details{
\code{db_temporary_table()} \strong{overwrites} an existing table
with name \code{to}.

\code{db_range_join()} \strong{overwites} and existing table with name \code{to}.

The table \code{key} is usually \code{"hg19"} or \code{"hg38"} and must have
\verb{#CHROM} and \code{POS} columns. The table \code{join} must have columns
\verb{#CHROM}, \code{start} and \code{end}. Following \emph{Bioconductor} convention
and as reported in at \code{am_browse()}, coordinates are 1-based and
ranges defined by \code{start} and \code{end} are closed. All columns from
both \code{key} and \code{join} are included, so column names (other than
\verb{#CHROM}) cannot be duplicated.
}
\examples{
db <- am_data("hg38")
db_tables(db)

## ranges of interest -- the first 200000 bases on chromsomes 1-4.
ranges <- tibble(
    `#CHROM` = paste0("chr", 1:4),
    start = rep(1, 4),
    end = rep(200000, 4)
)
db_temporary_table(db, ranges, "ranges")

db_tables(db)

rng <- db_range_join(db, "hg38", "ranges", "ranges_overlaps")
rng
rng |>
    count(`#CHROM`) |>
    arrange(`#CHROM`)

db_disconnect(db)

}
