% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_clinvar.R
\name{integrate_clinvar}
\alias{integrate_clinvar}
\title{Integrate ClinVar Pathogenicity Scores with AlphaMissense Annotations}
\usage{
integrate_clinvar(protein, clinvar_data)
}
\arguments{
\item{protein}{a valid UniProt accession identifier.}

\item{am_table}{a tibble derived from \code{am_data("aa_substitution")}.
Alternatively, a user-defined tibble or dataframe with columns
\code{uniprot_id}, \code{protein_variant}, \code{am_class}, and \code{am_pathogenicity}.
The columns are as follow:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ - `uniprot_id`: UniProt accession identifier(s).
 - `protein_variant`: variant identifier string, with the protein.
 position in the middle and the reference and mutant amino acid residues
 to the left and right of the position, respectively.
 - `am_class`: AlphaMissense classification of either "benign",
 "ambiguous", or "pathogenic".
 - `am_pathogenicity`: AlphaMissense predicted score.
}\if{html}{\out{</div>}}}

\item{cv_table}{a tibble or dataframe containing ClinVar information. By
default, ClinVar information is derived from the supplemental table of
the \href{https://www.science.org/doi/10.1126/science.adg7492}{2023}
AlphaMissense paper. Alternatively, a user-defined tibble or dataframe
with columns \code{accession} containing UniProt identifiers,
\code{variant_id} matching protein variants in AlphaMissense, and \code{label}
of binary values 0 and 1 for benign or pathogenic, respectively.}
}
\value{
\code{integrate_clinvar()} returns a \code{ggplot} object which overlays
ClinVar pathogenicity annotations onto AlphaMissense predicted scores
for comparison and visualization.
}
\description{
\code{integrate_clinvar()} integrates ClinVar pathogenicity scores
with AlphaMissense predicted scores derived from
\code{am_data("aa_substitutions")} and returns a ggplot object for
visualization.
}
\examples{
integrate_clinvar(protein = "P37023")

}
