% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clinvar.R
\name{plot_clinvar}
\alias{plot_clinvar}
\title{Integrate ClinVar Classifications with AlphaMissense Pathogenicity Scores}
\usage{
plot_clinvar(protein, am_table, cv_table = clinvar_data)
}
\arguments{
\item{am_table}{a data.frame #link to description - derived from \code{am_data("aa_substitution")}.
Alternatively, a user-defined tibble or dataframe.
Columns must include:
\itemize{
\item \code{uniprot_id}: UniProt accession identifier(s).
\item \code{protein_variant}: variant identifier string, with the protein.
position in the middle and the reference and mutant amino acid residues
to the left and right of the position, respectively.
\item \code{am_class}: AlphaMissense classification of either "benign",
"ambiguous", or "pathogenic".
\item \code{am_pathogenicity}: AlphaMissense predicted score.
}}

\item{cv_table}{a tibble or dataframe containing ClinVar information. By
default, ClinVar information is derived from the supplemental table of
the AlphaMissense
\href{https://www.science.org/doi/10.1126/science.adg7492}{[2023]} paper.
Alternatively, a user-defined tibble or dataframe.
Columns must include:
\itemize{
\item \code{uniprot_id}: UniProt accession identifier(s), matching AlphaMissense table.
\item \code{protein_variant}: protein variant identifier string, matching AlphaMissense
format.
\item \code{cv_class}: binary values 0 (benign) and 1 (pathogenic) in ClinVar.
}}

\item{pID}{a valid UniProt accession identifier. # check Martin's code}
}
\value{
\code{integrate_clinvar()} returns a \code{ggplot} object which overlays
ClinVar pathogenicity annotations onto AlphaMissense predicted scores
for comparison and visualization.
}
\description{
\code{plot_clinvar()} integrates ClinVar classifications
with AlphaMissense predicted scores derived from
\code{am_data("aa_substitutions")} and returns a ggplot object for
visualization.
}
\examples{
data(clinvar_data)

# get AM table with that pID first

plot_clinvar(protein = "P37023", cv_table = clinvar_data, am_table = X)

}
