% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_granges.R
\name{plot_granges}
\alias{plot_granges}
\title{Plot a GRanges Object Using Shiny and Gosling}
\usage{
plot_granges(
  gr_input,
  title = "GRanges Plot",
  subtitle = "Stacked nucleotide example",
  plot_type = c("bars", "lollipop")
)
}
\arguments{
\item{title}{character(1) Title of the plot. Default is
"GRanges Plot".}

\item{subtitle}{character(1) The subtitle of the plot. Default is
"Stacked nucleotide example".}

\item{plot_type}{character(1) Select the type of gosling plot.
Default is "bars"; available types are described in Details.}

\item{gr}{A \code{\link{GRanges}} object containing the genomic ranges to be
plotted.}
}
\value{
A \code{shinyApp} object that, when run, displays the Gosling
plot.
}
\description{
\code{plot_granges()} creates a Shiny app that displays a
Gosling plot of a given GRanges object.  It visualizes genomic
ranges with both rectangle and point representations, and
allows for customization of the plot title and subtitle.
}
\details{
The function supports 2 types of plots as selected through the
\code{plot_type} argument: (1) \code{"bars"}: a barplot-like view of the
pathogenicity score at each position, similar to a sequencing
coverage plot, and (2) \code{"lollipop"}: a lollipop plot focusing on
the pathogenicity classification (ambiguous, benign, pathogenic) at
each position. It requires that the \code{\link{GRanges}} object has the
following metadata columns: 'am_class' (effect
classification),'am_pathogenicity' (pathogenicity score), 'ALT'
(alternative allele) and 'REF' (reference allele).
}
\note{
This function requires the \code{shiny}, \code{shiny.gosling}, and
\code{GenomicRanges} packages to be installed.
}
\examples{
if (requireNamespace("GenomicRanges")) {

   ## Create a sample GRanges object from AlphamissenseR
   gpos <-
       am_data("hg38") |>
       filter(uniprot_id == "Q1W6H9") |>
       to_GPos()

   ## Plot the GRanges object
   plot_granges(
       gpos, title = "Q1W6H_track", subtitle = "bar plot example",
       plot_type = "bars"
   )

}

}
