% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shinygosling.R
\name{plot_granges}
\alias{plot_granges}
\title{Plot a GRanges object using Shiny and Gosling}
\usage{
plot_granges(
  gr,
  title = "GRanges Plot",
  subtitle = "Stacked nucleotide example",
  plot_type = "bars"
)
}
\arguments{
\item{gr}{A GRanges object containing the genomic ranges to
be plotted.}

\item{title}{Character string. The title of the plot. Default is
"GRanges Plot".}

\item{subtitle}{Character string. The subtitle of the plot. Default
is "Stacked nucleotide example".}

\item{plot_type}{Character string. Select the type of gosling plot.
Default is "bar"
- "bars": Stacked bar plot with height based on pathogenicity score
- "lollipop": variation of a bar chart where the bar is replaced with a
line and a dot at the end to show mutation variations.}
}
\value{
A Shiny app object that, when run, displays the Gosling
plot.
}
\description{
This function creates a Shiny app that displays a
Gosling plot of a given GRanges object.  It visualizes genomic
ranges with both rectangle and point representations, and
allows for customization of the plot title and subtitle.
The function creates two tracks: a reference track with
rectangles and an alternative track with points. It assumes
that the GRanges object has 'am_class' and 'ALT' metadata
columns. The 'am_class' column is used for coloring the points,
while 'ALT' is used for the text of the points.
}
\note{
This function requires the shiny, shiny.gosling, and GenomicRanges
packages to be installed.
}
\examples{
if (requireNamespace("GenomicRanges")) {

## Create a sample GRanges object from AlphamissenseR
gpos <-
   am_data("hg38") |>
   filter(uniprot_id == "Q1W6H9") |>
   to_GPos()

## Plot the GRanges object
plot_granges(gpos, mode = "bar", title = "Q1W6H9 track", subtitle 
= "bar plot example")
}


}
