% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinGym.R
\name{ProteinGym_AlphaMissense_data}
\alias{ProteinGym_AlphaMissense_data}
\alias{ProteinGym_DMS_data}
\alias{ProteinGym_correlation_plot}
\title{Integrate ProteinGym DMS and AlphaMissense Pathogenicity Scores}
\usage{
ProteinGym_AlphaMissense_data()

ProteinGym_DMS_data()

ProteinGym_correlation_plot(uniprotId, alphamissense_table, DMS_table)
}
\arguments{
\item{uniprotId}{\code{character()} a valid UniProt accession identifier.}

\item{alphamissense_table}{a table containing AlphaMissense predictions
for variants matching ProteinGym substitution mutants. The default
table is derived from the supplemental data of the AlphaMissense paper.
Alternatively, a user-defined \code{\link[tibble:tbl_df-class]{tibble::tbl_df}} or \code{\link{data.frame}}
can be supplied.}

\item{DMS_table}{a table containing deep mutational scanning (DMS)
assay scores for protein variants matching AlphaMissense. The default
table loads in substitutions from
\href{https://proteingym.org/download}{ProteinGym}.
Alternatively, a user-defined \code{\link[tibble:tbl_df-class]{tibble::tbl_df}} or \code{\link{data.frame}}
can be supplied.}
}
\value{
\code{ProteinGym_AlphaMissense_data()} returns a data.frame with 1622429 rows and
4 variables:
\itemize{
\item \code{DMS_id}: ProteinGym assay identifier.
\item \code{Uniprot_ID}: UniProt accession identifier.
\item \code{variant_id}: Mutant identifier string matching ProteinGym.
Protein position in the middle, and the reference and mutant
amino acid residues to the left and right of the position, respectively.
\item \code{AlphaMissense}: AlphaMissense pathogenicity score.
}

\code{ProteinGym_DMS_data()} returns a list of 216 data.frames corresponding to
individual DMS assays. Each table contains the following 6 columns:
\itemize{
\item \code{UniProt_id}: UniProt accession identifier.
\item \code{DMS_id}: ProteinGym assay identifier.
\item \code{mutant}: Mutant identifier string matching AlphaMissense variants.
Specifically, the set of substitutions to apply on the reference sequence
to obtain the mutated sequence (e.g., A1P:D2N implies the amino acid 'A'
at position 1 should be replaced by 'P', and 'D' at position 2 should be
replaced by 'N').
\item \code{mutated_sequence}: Full amino acid sequence for the mutated protein.
\item \code{DMS_score}: Experimental measurement in the DMS assay.
Higher values indicate higher fitness of the mutated protein.
\item \code{DMS_score_bin}: Factor, indicating whether the DMS_score is
above the fitness cutoff (1 is fit, 0 is not fit).
}

\code{ProteinGym_correlation_plot()} returns a \code{ggplot} object visualizing the
Spearman correlation between experimental DMS scores and AlphaMissense
predicted scores. In our case, a stronger negative correlation correspond
to a stronger relationship between the two measures.
}
\description{
\code{ProteinGym_AlphaMissense_data()} loads in the AlphaMissense
pathogenicity scores for mutants in ProteinGym from the AlphaMissense
publication by Cheng et al.
(\href{https://www.science.org/doi/10.1126/science.adg7492}{2023}).

\code{ProteinGym_DMS_data()} loads in 216 ProteinGym deep mutational
scanning assays (DMS) for substitutions. The data is provided by Notin et. al
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC10723403/}{(2023)}.

\code{ProteinGym_correlation_plot()} runs a Spearman correlation between
ProteinGym deep mutational scanning (DMS) assay scores against
AlphaMissense predicted scores. Returns a ggplot object for visualization.
}
\details{
For \code{ProteinGym_correlation_plot()},
\code{alphamissense_table} columns must include:
\itemize{
\item \code{Uniprot_ID}: UniProt accession identifier.
\item \code{variant_id}: Mutant identifier string matching ProteinGym.
Protein position in the middle, and the reference and mutant
amino acid residues to the left and right of the position, respectively.
\item \code{AlphaMissense}: AlphaMissense pathogenicity score.
}

\code{DMS_table} columns must include:
\itemize{
\item \code{UniProt_id}: UniProt accession identifier.
\item \code{mutant}: Mutant identifier string matching AlphaMissense variants.
Specifically, the set of substitutions to apply on the reference sequence
to obtain the mutated sequence (e.g., A1P:D2N implies the amino acid 'A'
at position 1 should be replaced by 'P', and 'D' at position 2 should be
replaced by 'N').
\item \code{DMS_score}: Experimental measurement in the DMS assay.
Higher values indicate higher fitness of the mutated protein.
}
}
\examples{
am_table <- ProteinGym_AlphaMissense_data()

pg_data <- ProteinGym_DMS_data()


ProteinGym_correlation_plot(uniprotId = "Q9NV35")

}
