% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaMissense.R
\name{alphamissense_browse}
\alias{alphamissense_browse}
\alias{alphamissense_available}
\alias{alphamissense}
\title{Retrieve AlphaMissense Resources as DuckDB Databases}
\usage{
alphamissense_browse(record = ALPHAMISSENSE_RECORD)

alphamissense_available(record = ALPHAMISSENSE_RECORD)

alphamissense(key = alphamissense_available(), bfc = BiocFileCache())
}
\arguments{
\item{record}{character(1) Zenodo record for the AlphaMissense data
resources.}

\item{key}{a character(1) 'key' from the result of
\code{alphamissense_available()}, or a single row of the tibble
returned by \code{alphamissense_available()}.}

\item{bfc}{an object returned by \code{BiocFileCache()} representing the
location where downloaded files and the parsed database will be
stored. The default is the 'global' BiocFileCache.}
}
\value{
\code{alphamissense_available()} returns a tibble with columns
\code{key}, \code{size}, and \code{link}. The meaning of key must be
determined with reference to the information at
\code{alphamissense_browse()}.

\code{alphamissense()} returns a dbplyr (database) tibble
represented the downloaded and parsed file. Fields in the
database are as described on the Zenodo resource page.
}
\description{
\code{alphamissense_browse()} opens a web browser at the
Zenodo record for the AlphaMissense data.

\code{alphamissense_available()} reports available datasets
in the record.

\code{alphamissense()} retrieves a single \code{key} from the
the AlpahMissense Zenodo site and parses the file into a DuckDB
database.
}
\details{
\code{alphamissense()} uses BiocFileCache to store the
downloaded file and the corresponding DuckDB database.
}
\examples{
if (interactive())
    alphamissense_browse()

alphamissense_available()

alphamissense("AlphaMissense_hg38.tsv.gz")

}
