% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_clinvar.R
\name{integrate_clinvar}
\alias{integrate_clinvar}
\title{Integrate ClinVar Pathogenicity Scores with AlphaMissense Annotations}
\usage{
integrate_clinvar(protein, am_table, cv_table = clinvar_data)
}
\arguments{
\item{protein}{a valid UniProt accession identifier.}

\item{am_table}{a tibble derived from \code{am_data("aa_substitution")}.
Alternatively, a user-defined tibble or dataframe.
Columns must include:
\itemize{
\item \code{uniprot_id}: UniProt accession identifier(s).
\item \code{protein_variant}: variant identifier string, with the protein.
position in the middle and the reference and mutant amino acid residues
to the left and right of the position, respectively.
\item \code{am_class}: AlphaMissense classification of either "benign",
"ambiguous", or "pathogenic".
\item \code{am_pathogenicity}: AlphaMissense predicted score.
}}

\item{cv_table}{a tibble or dataframe containing ClinVar information. By
default, ClinVar information is derived from the supplemental table of
the AlphaMissense
\href{https://www.science.org/doi/10.1126/science.adg7492}{[2023]} paper.
Alternatively, a user-defined tibble or dataframe.
Columns must include:
\itemize{
\item \code{uniprot_id}: UniProt accession identifier(s), matching AlphaMissense table.
\item \code{protein_variant}: protein variant identifier string, matching AlphaMissense
format.
\item \code{cv_class}: binary values 0 (benign) and 1 (pathogenic) in ClinVar.
}}
}
\value{
\code{integrate_clinvar()} returns a \code{ggplot} object which overlays
ClinVar pathogenicity annotations onto AlphaMissense predicted scores
for comparison and visualization.
}
\description{
\code{integrate_clinvar()} integrates ClinVar pathogenicity scores
with AlphaMissense predicted scores derived from
\code{am_data("aa_substitutions")} and returns a ggplot object for
visualization.
}
\examples{
integrate_clinvar(protein = "P37023")

}
