% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clinvar.R
\name{plot_clinvar}
\alias{plot_clinvar}
\title{Integrate ClinVar Labels with AlphaMissense Pathogenicity Scores}
\usage{
plot_clinvar(uniprotId, am_table, cv_table)
}
\arguments{
\item{uniprotId}{a valid UniProt accession identifier}

\item{am_table}{a table derived from \code{am_data("aa_substitution")}.
Alternatively, a user-defined tibble or data.frame.
Columns must include:
\itemize{
\item \code{uniprot_id}: UniProt accession identifier(s).
\item \code{protein_variant}: variant identifier string, with the protein.
position in the middle and the reference and mutant amino acid residues
to the left and right of the position, respectively.
\item \code{am_class}: AlphaMissense classification of either "benign",
"ambiguous", or "pathogenic".
\item \code{am_pathogenicity}: AlphaMissense predicted score.
}}

\item{cv_table}{a table containing ClinVar information derived from the
supplemental table of the AlphaMissense
\href{https://www.science.org/doi/10.1126/science.adg7492}{[2023]} paper.
Alternatively, a user-defined tibble or data.frame.
Columns must include:
\itemize{
\item \code{uniprot_id}: UniProt accession identifier(s), matching \code{am_table}.
\item \code{protein_variant}: variant identifier string, matching \code{am_table} format.
\item \code{cv_class}: binary ClinVar classification of 0 (benign) or 1 (pathogenic).
}}
}
\value{
\code{plot_clinvar()} returns a \code{ggplot} object which overlays
ClinVar classifications onto AlphaMissense predicted scores for
comparison and visualization. Blue, gray, and red colors represent
pathogenicity classifications for "likely benign", "ambiguous", or
"likely pathogenic", respectively. Large, bolded points represent
ClinVar variants and are colored according to their true clinical
classification, while smaller dots in the background are AlphaMissense
predictions.
}
\description{
\code{plot_clinvar()} integrates ClinVar classifications
with AlphaMissense predicted scores derived from
\code{am_data("aa_substitutions")} and returns a ggplot object for
visualization.
}
\examples{
data(clinvar_data)
am_table <- db_connect() |>
            tbl("aa_substitutions") |>
            filter(uniprot_id == "P37023") |>
            dplyr::as_tibble()
plot_clinvar(uniprotId = "P37023",
                am_table = am_table,
                cv_table = clinvar_data)

}
