% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clinvar.R
\name{plot_clinvar}
\alias{plot_clinvar}
\title{Integrate ClinVar Labels with AlphaMissense Pathogenicity Scores}
\usage{
plot_clinvar(uniprotId, alphamissense_table, clinvar_table)
}
\arguments{
\item{uniprotId}{a string with a valid UniProt accession
identifier.}

\item{alphamissense_data}{a table containing AlphaMissense predictions for
protein variants. By default, the table is derived from
\code{am_data("aa_substitution")}. Alternatively, a user-defined
\code{\link[tibble:tibble]{tibble::tibble()}} or \code{\link[base:data.frame]{base::data.frame()}} can be supplied.}

\item{clinvar_data}{a table containing ClinVar information. By default, the
table is derived from the supplemental data of the AlphaMissense paper.
Alternatively, a user-defined \code{\link[tibble:tibble]{tibble::tibble()}} or \code{\link[base:data.frame]{base::data.frame()}}
can be supplied.}
}
\value{
\code{plot_clinvar()} returns a \code{ggplot} object which overlays ClinVar
classifications onto AlphaMissense predicted scores. Blue, gray, and red
colors represent pathogenicity classifications for "likely benign",
"ambiguous", or "likely pathogenic", respectively. Large, bolded points
are ClinVar variants colored according to their clinical classification,
while smaller points in the background are AlphaMissense predictions.
}
\description{
\code{plot_clinvar()} integrates ClinVar classifications with
AlphaMissense predicted scores derived from \code{am_data("aa_substitutions")}
and returns a ggplot object for visualization.
}
\details{
For \code{alphamissense_table}, columns must include:
\itemize{
\item \code{uniprot_id}: UniProt accession identifier.
\item \code{protein_variant}: variant identifier string, with protein
position in the middle, and the reference and mutant amino acid
residues to the left and right of the position, respectively.
\item \code{am_class}: AlphaMissense classification of either "benign",
"ambiguous", or "pathogenic".
\item \code{am_pathogenicity}: AlphaMissense predicted score.
}

For \code{clinvar_table}, columns must include:
\itemize{
\item \code{uniprot_id}: UniProt accession identifier, matching \code{alphamissense_table}.
\item \code{protein_variant}: variant identifier string, matching
\code{alphamissense_table} format.
\item \code{cv_class}: binary ClinVar classification of 0 (benign) or 1 (pathogenic).
}
}
\examples{

data(clinvar_data)

alphamissense_table <- am_data("aa_substitutions") |>
                        filter(uniprot_id == "P37023") |>
                        dplyr::collect()

plot_clinvar(uniprotId = "P37023",
             alphamissense_table = alphamissense_table,
             clinvar_table = clinvar_data)

}
\references{
Cheng et al.,
Accurate proteome-wide missense variant effect prediction with AlphaMissense.
\emph{Science} 381, eadg7492. DOI:10.1126/science.adg7492.
}
