% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shinygosling.R
\name{plot_granges}
\alias{plot_granges}
\title{Plot a GRanges object using Shiny and Gosling}
\usage{
plot_granges(
  as_GRanges,
  title = "GRanges Plot",
  subtitle = "Stacked nucleotide example"
)
}
\arguments{
\item{as_GRanges}{A GRanges object containing the genomic ranges to
be plotted.}

\item{title}{Character string. The title of the plot. Default is
"GRanges Plot".}

\item{subtitle}{Character string. The subtitle of the plot. Default
is "Stacked nucleotide example".}
}
\value{
A Shiny app object that, when run, displays the Gosling
plot.
}
\description{
This function creates a Shiny app that displays a
Gosling plot of a given GRanges object.  It visualizes genomic
ranges with both rectangle and point representations, and
allows for customization of the plot title and subtitle.
The function creates two tracks: a reference track with
rectangles and an alternative track with points. It assumes
that the GRanges object has 'am_class' and 'ALT' metadata
columns. The 'am_class' column is used for coloring the points,
while 'ALT' is used for the text of the points.
}
\note{
This function requires the shiny, gosling, and GenomicRanges
packages to be installed.
}
\examples{
\dontrun{
library(GenomicRanges)

# Create a sample GRanges object
gr <- GRanges(
  seqnames = c("chr1", "chr2", "chr1"),
  ranges = IRanges(start = c(1, 3, 5), end = c(2, 4, 6)),
  am_class = c("likely_benign", "ambiguous", "likely_pathogenic"),
  ALT = c("A", "T", "G")
)

# Plot the GRanges object
AlphaMissenseR::plot_granges(gr, title = "My GRanges Plot", subtitle = "Custom subtitle")
}

}
