% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinGym.R
\name{proteingym_data}
\alias{proteingym_data}
\title{Integrate ProteinGym Mave Scores with AlphaMissense Pathogenicity
Scores}
\usage{
proteingym_data(record = ALPHAMISSENSE_RECORD, bfc = BiocFileCache())
}
\arguments{
\item{record}{character(1) Zenodo record for the AlphaMissense data
resources.}

\item{bfc}{an object returned by \code{BiocFileCache()} representing the
location of the AlphaMissenseR database. The default is the
'global' BiocFileCache.}
}
\value{
\code{proteingym_data()} returns a tbl with 82872 rows and 5 variables:
\itemize{
\item \code{variant_id}: ClinVar variant identifier.
\item \code{transcript_id}: Ensembl transcript identifier.
\item \code{protein_variant}: UniProt accession:protein variant identifier.
\item \code{AlphaMissense}: AlphaMissense pathogenicity score.
\item \code{label}: Binary ClinVar class, either "benign" or "pathogenic".
}
}
\description{
\code{proteingym_data()} loads ProteinGym information from the
supplemental table of the AlphaMissense
\href{https://www.science.org/doi/10.1126/science.adg7492}{[2023]}
paper.
}
\examples{
proteingym_data()
}
